requirements.txt
selenium
pytest
requests
allure-pytest
ipdb
python-environ
setuptools
pytest-repeat


pre-commit
isort
black
pylama[toml]
